django-koldar-utils/requirements.txt
alabaster==0.7.12
aniso8601==8.1.1
arrow==1.1.0
asgiref==3.3.4
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
datasize==1.0.0
decorator==4.4.2
Django==3.2.7
django-appconf==1.0.4
django-currentuser==0.5.3
django-environ==0.4.5
django-filter==2.4.0
django-graphql-jwt==0.3.2
django-polymorphic==3.0.0
djangorestframework==3.12.4
docutils==0.17.1
gitdb==4.0.7
GitPython==3.1.18
graphene==3.0b7
graphene-django==3.0.0b7
graphene-django-extras==0.5.1
graphql-core==3.1.5
graphql-relay==3.1.0
idna==2.10
imagesize==1.2.0
inflect==5.3.0
inflection==0.5.1
Jinja2==3.0.1
jmespath==0.10.0
koldar-utils==2.103.0
m2r==0.2.1
MarkupSafe==2.0.1
mistune==0.8.4
networkx==2.5.1
packaging==21.0
Pillow==8.3.1
promise==2.3
pydot==1.4.2
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
Rx==1.6.1
semantic-version==2.8.5
singledispatch==3.6.1
six==1.16.0
smmap==4.0.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-autodoc-typehints==1.12.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
stringcase==1.2.0
text-unidecode==1.3
urllib3==1.26.5
